backend: Extract non-db migrations into seperate file
In preparation for further migrations, move the current CA file
migration into dedicated file.
